+++ src/base/optimise/owl_optimise_generic.ml
@@ -650,6 +650,7 @@ module Make
       Checkpoint.(state.gs <- gs');
       Checkpoint.(state.ps <- ps');
       Checkpoint.(state.current_batch <- state.current_batch + 1);
+      (** TODO: Gc.minor (); *)
     done;
 
     (* print optimisation summary *)