+++ src/owl/dense/owl_dense_ndarray_generic.mli
@@ -420,6 +420,8 @@ val repeat : ?axis:int -> ('a, 'b) t -> int -> ('a, 'b) t
 ``repeat ~axis x a`` repeats the elements along ``axis`` for ``a`` times. The default
 value of ``?axis`` is the highest dimension of ``x``. This function is similar to
 ``numpy.repeat`` except that ``a`` is an integer instead of an array.
+
+Not that the ``axis`` can be negative.
  *)
 
 val concat_vertical : ('a, 'b) t -> ('a, 'b) t -> ('a, 'b) t
@@ -1471,10 +1473,13 @@ val softsign : (float, 'a) t -> (float, 'a) t
 elements in ``x`` and returns the result in a new ndarray.
  *)
 
-val softmax : (float, 'a) t -> (float, 'a) t
+val softmax : ?axis:int -> (float, 'a) t -> (float, 'a) t
 (**
 ``softmax x`` computes the softmax functions ``(exp x) / (sum (exp x))`` of
-all the elements in ``x`` and returns the result in a new array.
+all the elements along the specified ``axis`` in ``x`` and returns the result
+in a new ndarray.
+
+By default, ``axis = -1``, i.e. along the highest dimension.
  *)
 
 val sigmoid : (float, 'a) t -> (float, 'a) t
@@ -2380,7 +2385,7 @@ val sigmoid_ : ('a, 'b) t -> unit
 ``sigmoid_ x`` is similar to ``sigmoid`` but output is written to ``x``
  *)
 
-val softmax_ : ('a, 'b) t -> unit
+val softmax_ : ?axis:int -> ('a, 'b) t -> unit
 (**
 ``softmax_ x`` is similar to ``softmax`` but output is written to ``x``
  *)